+++ CHANGELOG.rst
@@ -39,6 +39,7 @@ Fixed
 -----
 - Fixed ``probabilities_ket`` computation in C++ simulator. (#580)
 - Fixed bug in the definition of ``cswap`` gate and its test (#685).
+- Fixed the examples to be compatible with version 0.5+ (#672)
 
 
 `0.5.6`_ - 2018-07-06